Move the panel into individual namespace.
