+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/Panels/InvalidPanel.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Panels
 {
-    public class PropertyPanel : Panel
+    public class InvalidPanel : Panel
     {
-        public PropertyPanel()
+        public InvalidPanel()
         {
             Width = 200;
         }
