+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/Panels/PropertyPanel.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.Panels
 {
-    public class InvalidPanel : Panel
+    public class PropertyPanel : Panel
     {
-        public InvalidPanel()
+        public PropertyPanel()
         {
             Width = 200;
         }
